+++ test/test_internals/test_asyncredis.py
@@ -1,5 +1,3 @@
-import re
-
 import pytest
 import redis
 import redis.asyncio
@@ -14,16 +12,6 @@ pytestmark.extend(
         pytest.mark.asyncio,
     ]
 )
-
-
-@fake_only
-@testtools.run_test_if_redispy_ver("gte", "5.1")
-async def test_repr_redis_51(async_redis: redis.asyncio.Redis):
-    assert re.fullmatch(
-        r"<redis.asyncio.connection.ConnectionPool\("
-        r"<fakeredis.aioredis.FakeConnection\(server=<fakeredis._server.FakeServer object at .*>,db=0\)>\)>",
-        repr(async_redis.connection_pool),
-    )
 
 
 @fake_only