+++ test/test_internals/test_init_args.py
@@ -22,8 +22,9 @@ def test_deprecated_args_show_warnings_for_retry_on_timeout(warn_mock: mock.Magi
 @run_test_if_redispy_ver("gte", "6")
 @mock.patch("warnings.warn")
 def test_deprecated_args_no_warnings(warn_mock: mock.MagicMock):
-    fakeredis.FakeStrictRedis(host="localhost", port=6390, db=0)
+    fakeredis.FakeStrictRedis(host="localhost", port=6390, db=0, version=6)
     warn_mock.assert_not_called()
+    fakeredis.FakeAsyncRedis(host="localhost", port=6390, db=0, version=6)
 
 
 @pytest.mark.fake