+++ docs/about/changelog.md
@@ -7,6 +7,21 @@ tags:
 toc_depth: 2
 ---
 
+## v2.30.4 -
+
+### 🐛 Bug Fixes
+
+- version init argument typing #401
+- include `lib_name` and `lib_version` in `FakeBaseConnectionMixin`
+- when using `FakeValkey`, raise `valkey.ResponseError` instead of `redis.ResponseError` #402
+
+## v2.30.3 - 2025-08-01
+
+### 🐛 Bug Fixes
+
+- Import `Callable` from `typing` to support python 3.8 #398
+- FakeValkey init arguments #400
+
 ## v2.30.2 - 2025-08-01
 
 ### 🐛 Bug Fixes
@@ -23,7 +38,7 @@ toc_depth: 2
 ### 🐛 Bug Fixes
 
 - Move `LICENSE` file to fakeredis/ on build #395
-- Show warning for deprecated parameter only if parameter is included in instantiation #384
+- Show warning for deprecated parameter only if the parameter is included in instantiation #384
 
 ### 🧰 Maintenance
 
@@ -38,7 +53,7 @@ toc_depth: 2
 
 ### 🐛 Bug Fixes
 
-- aio.FakeRedis: Using `inspect` to get `__init__` paramaters #384
+- aio.FakeRedis: Using `inspect` to get `__init__` parameters #384
 - Discrepancy when using xread with resp3 #393
 
 ## v2.29.0 - 2025-05-01